+++ src/test/java/com/numble/instagram/presentation/follow/FollowControllerDocTest.java
@@ -1,8 +1,8 @@
 package com.numble.instagram.presentation.follow;
 
 import com.numble.instagram.application.auth.token.TokenProvider;
-import com.numble.instagram.application.usecase.CreateFollowUserUsecase;
-import com.numble.instagram.application.usecase.DestroyFollowUserUsecase;
+import com.numble.instagram.application.usecase.follow.CreateFollowUserUsecase;
+import com.numble.instagram.application.usecase.follow.DestroyFollowUserUsecase;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
